+++ src/Draco.Compiler/Internal/Symbols/Metadata/MetadataStaticClassSymbol.cs
@@ -22,6 +22,8 @@ internal sealed class MetadataStaticClassSymbol(
 
     public override string Name => this.MetadataName;
     public override string MetadataName => this.MetadataReader.GetString(typeDefinition.Name);
+    public MetadataReader MetadataReader => this.Assembly.MetadataReader;
+    public override Symbol ContainingSymbol { get; } = containingSymbol;
 
     public override Api.Semantics.Visibility Visibility => typeDefinition.Attributes.HasFlag(TypeAttributes.Public) ? Api.Semantics.Visibility.Public : Api.Semantics.Visibility.Internal;
 
@@ -31,22 +33,15 @@ internal sealed class MetadataStaticClassSymbol(
     internal override string RawDocumentation => LazyInitializer.EnsureInitialized(ref this.rawDocumentation, this.BuildRawDocumentation);
     private string? rawDocumentation;
 
-    public override Symbol ContainingSymbol { get; } = containingSymbol;
-
     // NOTE: thread-safety does not matter, same instance
     public MetadataAssemblySymbol Assembly => this.assembly ??= this.AncestorChain.OfType<MetadataAssemblySymbol>().First();
     private MetadataAssemblySymbol? assembly;
 
-    public MetadataReader MetadataReader => this.Assembly.MetadataReader;
-
     public string? DefaultMemberAttributeName =>
         InterlockedUtils.InitializeMaybeNull(ref this.defaultMemberAttributeName, () => MetadataSymbol.GetDefaultMemberAttributeName(typeDefinition, this.Assembly.Compilation, this.MetadataReader));
-
-    public IEnumerable<Symbol> AdditionalSymbols =>
-        InterlockedUtils.InitializeDefault(ref this.additionalSymbols, this.BuildAdditionalSymbols);
-    private ImmutableArray<Symbol> additionalSymbols;
-
     private string? defaultMemberAttributeName;
+
+    public IEnumerable<Symbol> AdditionalSymbols => [];
 
     private ImmutableArray<Symbol> BuildMembers()
     {
@@ -118,8 +113,5 @@ internal sealed class MetadataStaticClassSymbol(
         XmlDocumentationExtractor.Extract(this);
 
     private string BuildRawDocumentation() =>
-        MetadataSymbol.GetDocumentation(this);
-
-    private ImmutableArray<Symbol> BuildAdditionalSymbols() =>
-        MetadataSymbol.GetAdditionalSymbols(this, typeDefinition, this.MetadataReader).ToImmutableArray();
+        MetadataDocumentation.GetDocumentation(this);
 }
